requirements.txt

djangorestframework
sentry-sdk
pillow